# schemahcl: speedup eval schemas