```
benchmark                old ns/op     new ns/op     delta
Benchmark_EvalHCL-16     196478424     170294770     -13.33%

benchmark                old allocs     new allocs     delta
Benchmark_EvalHCL-16     1069863        926489         -13.40%

benchmark                old bytes     new bytes     delta
Benchmark_EvalHCL-16     133138506     122031348     -8.34%
```